refactor(model): move relationship to common
